requirements.txt
transformers
shap
fastapi
uvicorn
pytorch
torchvision
pytest
python-dotenv